First round of cleanup on thread.rs.
