+++ src/bin/init/src/main.rs
@@ -385,7 +385,7 @@ fn main() {
                     unsafe {
                         let val = tr.as_ref().unwrap().wait(None);
                         if let Some(val) = val {
-                            if val != 0 {
+                            if val.0 != ExecutionState::Running {
                                 test_failed = true;
                             }
                         }
@@ -468,6 +468,6 @@ use twizzler_abi::{
         NewHandleFlags, ObjectCreate, ObjectCreateFlags, ThreadSync, ThreadSyncFlags, ThreadSyncOp,
         ThreadSyncReference, ThreadSyncSleep, ThreadSyncWake,
     },
-    thread::ThreadRepr,
+    thread::{ExecutionState, ThreadRepr},
 };
 use twizzler_object::{CreateSpec, Object, ObjectInitFlags};
