+++ src/kernel/src/syscall/thread.rs
@@ -9,7 +9,7 @@ use crate::{
 };
 
 pub fn sys_spawn(args: &ThreadSpawnArgs) -> Result<ObjID, ThreadSpawnError> {
-    crate::thread::start_new_user(*args)
+    crate::thread::entry::start_new_user(*args)
 }
 
 pub fn thread_ctrl(cmd: ThreadControl, arg: u64) -> (u64, u64) {
