+++ src/kernel/src/thread/state.rs
@@ -1,0 +1,9 @@
+#[derive(PartialEq, Copy, Clone, Debug)]
+#[repr(u32)]
+pub enum ThreadState {
+    Starting,
+    Running,
+    Blocked,
+    Exiting,
+    Exited,
+}
